requirements.txt
antlr4-python3-runtime==4.9.3
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
exceptiongroup==1.2.2
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
future==1.0.0
h5py==3.12.1
huggingface-hub==0.25.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
protobuf==5.28.2
pycparser==2.22
pyloudnorm==0.1.1
pyparsing==3.1.4
pypesq==1.2.4
pysptk==1.0.1
pytest==8.3.3
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pyworld==0.3.4
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.3
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.20.0
tomli==2.0.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchprofile==0.0.4
torchvision==0.13.1+cu113
tqdm==4.66.5
transformers==4.45.1
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
